계층 구성도 작성
