+++ server/src/main/java/com/example/tset/board/dto/BoardPatchDto.java
@@ -1,0 +1,18 @@
+package com.example.tset.board.dto;
+
+
+import lombok.Getter;
+
+@Getter
+public class BoardPatchDto {
+
+    private long broadId;
+
+    private String title;
+
+    private String content;
+
+    public void setBroadId(long broadId) {
+        this.broadId = broadId;
+    }
+}
