+++ server/src/main/java/com/example/tset/board/dto/BoardPostDto.java
@@ -1,16 +1,14 @@
 package com.example.tset.board.dto;
 
 
-import lombok.Builder;
 import lombok.Getter;
 
-import java.time.LocalDateTime;
-
-@Builder
 @Getter
-public class BroadReponseDto {
-    private long broadId;
+public class BoardPostDto {
+
     private String title;
+
     private String content;
+
 //    private LocalDateTime localDateTime;
 }
