--- server/src/main/java/com/example/tset/board/dto/BroadPostDto.java
@@ -1,16 +1,0 @@
-package com.example.tset.board.dto;
-
-
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-public class BroadPostDto {
-
-    private String title;
-
-    private String content;
-
-//    private LocalDateTime localDateTime;
-}
