+++ server/src/main/java/com/example/tset/board/mapper/BoardMapper.java
@@ -1,0 +1,4 @@
+package com.example.tset.board.mapper;
+
+public interface BoardMapper {
+}
