+++ server/src/main/java/com/example/tset/board/service/BoardService.java
@@ -1,0 +1,8 @@
+package com.example.tset.board.service;
+
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class BoardService {
+}
